server file restructure
